requirement.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-DotEnv==0.1.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
